feat: export link from index file
